day_26/requirements.txt
langchain-ollama>=0.1.0
langchain>=1.0.0
python-dotenv>=1.0.0